+++ elide-core/src/main/java/com/yahoo/elide/core/EntityBinding.java
@@ -32,6 +32,7 @@ import org.apache.commons.collections4.multimap.HashSetValuedHashMap;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
 
+import javax.persistence.AccessType;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Id;
@@ -43,6 +44,7 @@ import javax.persistence.Transient;
 import java.lang.annotation.Annotation;
 import java.lang.reflect.AccessibleObject;
 import java.lang.reflect.Field;
+import java.lang.reflect.Member;
 import java.lang.reflect.Method;
 import java.lang.reflect.Modifier;
 import java.util.ArrayList;
@@ -53,6 +55,7 @@ import java.util.Deque;
 import java.util.List;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentLinkedDeque;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -71,10 +74,17 @@ public class EntityBinding {
     public final Class<?> entityClass;
     public final String jsonApiType;
     public final String entityName;
-    @Getter private AccessibleObject idField;
-    @Getter private String idFieldName;
-    @Getter private Class<?> idType;
-    @Getter @Setter private Initializer initializer;
+    @Getter
+    private AccessibleObject idField;
+    @Getter
+    private String idFieldName;
+    @Getter
+    private Class<?> idType;
+    @Getter
+    @Setter
+    private Initializer initializer;
+    @Getter
+    private AccessType accessType;
 
     public final EntityPermissions entityPermissions;
     public final List<String> attributes;
@@ -88,6 +98,7 @@ public class EntityBinding {
     public final ConcurrentHashMap<String, CascadeType[]> relationshipToCascadeTypes = new ConcurrentHashMap<>();
     public final ConcurrentHashMap<String, AccessibleObject> fieldsToValues = new ConcurrentHashMap<>();
     public final MultiValuedMap<Pair<Class, String>, LifeCycleHook> fieldsToTriggers = new HashSetValuedHashMap<>();
+    public final MultiValuedMap<Class, LifeCycleHook> classToTriggers = new HashSetValuedHashMap<>();
     public final ConcurrentHashMap<String, Class<?>> fieldsToTypes = new ConcurrentHashMap<>();
     public final ConcurrentHashMap<String, String> aliasesToFields = new ConcurrentHashMap<>();
     public final ConcurrentHashMap<Method, Boolean> requestScopeableMethods = new ConcurrentHashMap<>();
@@ -95,6 +106,7 @@ public class EntityBinding {
     public final ConcurrentHashMap<Class<? extends Annotation>, Annotation> annotations = new ConcurrentHashMap<>();
 
     public static final EntityBinding EMPTY_BINDING = new EntityBinding();
+    private static final String ALL_FIELDS = "*";
 
     /* empty binding constructor */
     private EntityBinding() {
@@ -113,32 +125,89 @@ public class EntityBinding {
         entityClass = cls;
         jsonApiType = type;
         entityName = name;
+        inheritedTypes = getInheritedTypes(cls);
 
         // Map id's, attributes, and relationships
-        List<AccessibleObject> fieldOrMethodList = new ArrayList<>();
-        fieldOrMethodList.addAll(Arrays.asList(cls.getFields())
-                .stream()
-                .filter((field) -> ! Modifier.isStatic(field.getModifiers()))
-                .collect(Collectors.toList()));
+        List<AccessibleObject> fieldOrMethodList = getAllFields();
 
-        fieldOrMethodList.addAll(Arrays.asList(cls.getMethods())
-                .stream()
-                .filter((method) -> ! Modifier.isStatic(method.getModifiers()))
-                .collect(Collectors.toList()));
+        if (fieldOrMethodList.stream().anyMatch(field -> field.isAnnotationPresent(Id.class))) {
+            accessType = AccessType.FIELD;
+
+            /* Add all public methods that are computed */
+            fieldOrMethodList.addAll(
+                    getInstanceMembers(cls.getMethods(),
+                            (method) -> method.isAnnotationPresent(ComputedAttribute.class)
+                                    || method.isAnnotationPresent(ComputedRelationship.class))
+            );
+
+            //Elide needs to manipulate private fields that are exposed.
+            fieldOrMethodList.forEach(field -> field.setAccessible(true));
+        } else {
+            /* Preserve the behavior of Elide 4.2.6 and earlier */
+            accessType = AccessType.PROPERTY;
+
+            fieldOrMethodList.clear();
+
+            /* Add all public fields */
+            fieldOrMethodList.addAll(getInstanceMembers(cls.getFields()));
+
+            /* Add all public methods */
+            fieldOrMethodList.addAll(getInstanceMembers(cls.getMethods()));
+        }
 
         bindEntityFields(cls, type, fieldOrMethodList);
 
         attributes = dequeToList(attributesDeque);
         relationships = dequeToList(relationshipsDeque);
-        inheritedTypes = getInheritedTypes(cls);
         entityPermissions = new EntityPermissions(dictionary, cls, fieldOrMethodList);
+    }
+
+    /**
+     * Filters a list of class Members to instance methods & fields
+     *
+     * @param objects
+     * @param <T>
+     * @return A list of the filtered members
+     */
+    private <T extends Member> List<T> getInstanceMembers(T[] objects) {
+        return getInstanceMembers(objects, o -> true);
+    }
+
+    /**
+     * Filters a list of class Members to instance methods & fields
+     *
+     * @param objects    The list of Members to filter
+     * @param <T>        Concrete Member Type
+     * @param filteredBy An additional filter predicate to apply
+     * @return A list of the filtered members
+     */
+    private <T extends Member> List<T> getInstanceMembers(T[] objects, Predicate<T> filteredBy) {
+        return Arrays.stream(objects)
+                .filter(o -> !Modifier.isStatic(o.getModifiers()))
+                .filter(filteredBy)
+                .collect(Collectors.toList());
+    }
+
+    /**
+     * Get all fields of the entity class, including fields of superclasses (excluding Object)
+     * @return All fields of the EntityBindings entity class and all superclasses (excluding Object)
+     */
+    public List<AccessibleObject> getAllFields() {
+        List<AccessibleObject> fields = new ArrayList<>();
+
+        fields.addAll(getInstanceMembers(entityClass.getDeclaredFields(), (field) -> !field.isSynthetic()));
+        for (Class<?> type : inheritedTypes) {
+            fields.addAll(getInstanceMembers(type.getDeclaredFields(), (field) -> !field.isSynthetic()));
+        }
+
+        return fields;
     }
 
     /**
      * Bind fields of an entity including the Id field, attributes, and relationships.
      *
-     * @param cls Class type to bind fields
-     * @param type JSON API type identifier
+     * @param cls               Class type to bind fields
+     * @param type              JSON API type identifier
      * @param fieldOrMethodList List of fields and methods on entity
      */
     private void bindEntityFields(Class<?> cls, String type, Collection<AccessibleObject> fieldOrMethodList) {
@@ -182,8 +251,8 @@ public class EntityBinding {
     /**
      * Bind an id field to an entity.
      *
-     * @param cls Class type to bind fields
-     * @param type JSON API type identifier
+     * @param cls           Class type to bind fields
+     * @param type          JSON API type identifier
      * @param fieldOrMethod Field or method to bind
      */
     private void bindEntityId(Class<?> cls, String type, AccessibleObject fieldOrMethod) {
@@ -321,7 +390,7 @@ public class EntityBinding {
     /**
      * Check whether or not method expects a RequestScope.
      *
-     * @param method  Method to check
+     * @param method Method to check
      * @return True if accepts a RequestScope, false otherwise
      */
     public static boolean isRequestScopeableMethod(Method method) {
@@ -332,7 +401,7 @@ public class EntityBinding {
     /**
      * Check whether or not the provided method described a computed attribute or relationship.
      *
-     * @param method  Method to check
+     * @param method Method to check
      * @return True if method is a computed type, false otherwise
      */
     public static boolean isComputedMethod(Method method) {
@@ -367,7 +436,7 @@ public class EntityBinding {
             Method method = (Method) fieldOrMethod;
 
             int paramCount = method.getParameterCount();
-            Class<?> [] paramTypes = method.getParameterTypes();
+            Class<?>[] paramTypes = method.getParameterTypes();
 
             LifeCycleHook callback = (entity, scope, changes) -> {
                 try {
@@ -386,7 +455,11 @@ public class EntityBinding {
                 }
             };
 
-            bindTrigger(annotationClass, value, callback);
+            if (value.equals(ALL_FIELDS)) {
+                bindTrigger(annotationClass, callback);
+            } else {
+                bindTrigger(annotationClass, value, callback);
+            }
         }
     }
 
@@ -396,8 +469,19 @@ public class EntityBinding {
         fieldsToTriggers.put(Pair.of(annotationClass, fieldOrMethodName), callback);
     }
 
+    public void bindTrigger(Class<? extends Annotation> annotationClass,
+                            LifeCycleHook callback) {
+        classToTriggers.put(annotationClass, callback);
+    }
+
+
     public <A extends Annotation> Collection<LifeCycleHook> getTriggers(Class<A> annotationClass, String fieldName) {
         Collection<LifeCycleHook> methods = fieldsToTriggers.get(Pair.of(annotationClass, fieldName));
+        return methods == null ? Collections.emptyList() : methods;
+    }
+
+    public <A extends Annotation> Collection<LifeCycleHook> getTriggers(Class<A> annotationClass) {
+        Collection<LifeCycleHook> methods = classToTriggers.get(annotationClass);
         return methods == null ? Collections.emptyList() : methods;
     }
 
@@ -415,7 +499,7 @@ public class EntityBinding {
      * Return annotation from class, parents or package.
      *
      * @param annotationClass the annotation class
-     * @param <A> annotation type
+     * @param <A>             annotation type
      * @return the annotation
      */
     public <A extends Annotation> A getAnnotation(Class<A> annotationClass) {
@@ -433,7 +517,7 @@ public class EntityBinding {
     private List<Class<?>> getInheritedTypes(Class<?> entityClass) {
         ArrayList<Class<?>> results = new ArrayList<>();
 
-        for (Class<?> cls = entityClass.getSuperclass() ; cls != Object.class ; cls = cls.getSuperclass()) {
+        for (Class<?> cls = entityClass.getSuperclass(); cls != Object.class; cls = cls.getSuperclass()) {
             results.add(cls);
         }