+++ elide-core/src/main/java/com/yahoo/elide/core/filter/dialect/RSQLFilterDialect.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016, Yahoo Inc.
+ * Copyright 2018, Oath Inc.
  * Licensed under the Apache License, Version 2.0
  * See LICENSE file in project root for terms.
  */
@@ -15,6 +15,7 @@ import com.yahoo.elide.core.filter.expression.AndFilterExpression;
 import com.yahoo.elide.core.filter.expression.FilterExpression;
 import com.yahoo.elide.core.filter.expression.NotFilterExpression;
 import com.yahoo.elide.core.filter.expression.OrFilterExpression;
+import com.yahoo.elide.parsers.JsonApiParser;
 import com.yahoo.elide.utils.coerce.CoerceUtil;
 import cz.jirutka.rsql.parser.RSQLParser;
 import cz.jirutka.rsql.parser.RSQLParserException;
@@ -27,14 +28,7 @@ import cz.jirutka.rsql.parser.ast.RSQLOperators;
 import cz.jirutka.rsql.parser.ast.RSQLVisitor;
 
 import javax.ws.rs.core.MultivaluedMap;
-import java.io.File;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
@@ -60,10 +54,16 @@ public class RSQLFilterDialect implements SubqueryFilterDialect, JoinFilterDiale
 
     private final RSQLParser parser;
     private final EntityDictionary dictionary;
+    private final CaseSensitivityStrategy caseSensitivityStrategy;
 
     public RSQLFilterDialect(EntityDictionary dictionary) {
+        this(dictionary, new CaseSensitivityStrategy.FIQLCompliant());
+    }
+
+    public RSQLFilterDialect(EntityDictionary dictionary, CaseSensitivityStrategy caseSensitivityStrategy) {
         parser = new RSQLParser(getDefaultOperatorsWithIsnull());
-        this. dictionary = dictionary;
+        this.dictionary = dictionary;
+        this.caseSensitivityStrategy = caseSensitivityStrategy;
     }
 
     //add rsql isnull op to the default ops
@@ -97,11 +97,7 @@ public class RSQLFilterDialect implements SubqueryFilterDialect, JoinFilterDiale
         /*
          * Extract the last collection in the URL.
          */
-        String normalizedPath = Paths.get(path).normalize().toString().replace(File.separatorChar, '/');
-        if (normalizedPath.startsWith("/")) {
-            normalizedPath = normalizedPath.substring(1);
-        }
-
+        String normalizedPath = JsonApiParser.normalizePath(path);
         String[] pathComponents = normalizedPath.split("/");
         String lastPathComponent = pathComponents.length > 0 ? pathComponents[pathComponents.length - 1] : "";
 
@@ -295,22 +291,25 @@ public class RSQLFilterDialect implements SubqueryFilterDialect, JoinFilterDiale
             boolean endsWith = argument.endsWith("*");
             if (startsWith && endsWith && argument.length() > 2) {
                 String value = argument.substring(1, argument.length() - 1);
-                return new FilterPredicate(path, Operator.INFIX_CASE_INSENSITIVE, Collections.singletonList(value));
+                return new FilterPredicate(path, caseSensitivityStrategy.mapOperator(Operator.INFIX),
+                        Collections.singletonList(value));
             }
             if (startsWith && argument.length() > 1) {
                 String value = argument.substring(1, argument.length());
-                return new FilterPredicate(path, Operator.POSTFIX_CASE_INSENSITIVE, Collections.singletonList(value));
+                return new FilterPredicate(path, caseSensitivityStrategy.mapOperator(Operator.POSTFIX),
+                        Collections.singletonList(value));
             }
             if (endsWith && argument.length() > 1) {
                 String value = argument.substring(0, argument.length() - 1);
-                return new FilterPredicate(path, Operator.PREFIX_CASE_INSENSITIVE, Collections.singletonList(value));
+                return new FilterPredicate(path, caseSensitivityStrategy.mapOperator(Operator.PREFIX),
+                        Collections.singletonList(value));
             }
 
             Boolean isStringLike = path.lastElement()
                     .map(e -> e.getFieldType().isAssignableFrom(String.class))
                     .orElse(false);
             if (isStringLike) {
-                return new FilterPredicate(path, Operator.IN_INSENSITIVE, values);
+                return new FilterPredicate(path, caseSensitivityStrategy.mapOperator(Operator.IN), values);
             }
 
             return new FilterPredicate(path, Operator.IN, values);
@@ -318,7 +317,7 @@ public class RSQLFilterDialect implements SubqueryFilterDialect, JoinFilterDiale
 
         /**
          * Returns Predicate for '=isnull=' case depending on its arguments.
-         *
+         * <p>
          * NOTE: Filter Expression builder specially for '=isnull=' case.
          *
          * @return Returns Predicate for '=isnull=' case depending on its arguments.